add create, update and delete methods and single object view for clients

+++ cli/controllers/clients/__init__.py
@@ -1,3 +1,4 @@
+import json
 from typing import Optional
 from email_validator.exceptions_types import EmailNotValidError
 from email_validator.validate_email import validate_email
@@ -6,16 +7,14 @@ import typer
 import requests
 import typer
 from cli.helpers import authenticate, handle_response, sanitize_fullname
-from cli.views.clients import clients_list_view
+from cli.views.clients import clients_list_view, client_show_view
 from cli.views.shared import message_show_view
-from cli.controllers.users import list as users_list
 
 app = typer.Typer()
 
 
 @app.command()
 def create(
-    ctx: typer.Context,
     fullname: Annotated[
         str, typer.Option("--fullname", "-n", prompt=True, help="The client full name")
     ],
@@ -28,14 +27,6 @@ def create(
     company: Annotated[
         str, typer.Option("--company", "-c", prompt=True, help="The company name")
     ],
-    sales_contact: Annotated[
-        Optional[int],
-        typer.Option(
-            "--sales-contact",
-            "-s",
-            help="The id of the client's sales representant",
-        ),
-    ] = None,
 ):
     fullname = sanitize_fullname(fullname)
     try:
@@ -46,12 +37,24 @@ def create(
     except EmailNotValidError as e:
         message_show_view({"Error": "Email format is wrong", "Details": str(e)})
         raise typer.Exit()
-    authenticate()
-    if not sales_contact:
-        ctx.invoke(users_list)
-        sales_contact = typer.prompt("Select a sales representant")
-    print(sales_contact)
-    print("Creating Client")
+    new_client = {
+        "fullname": fullname,
+        "email": email,
+        "phone": phone,
+        "company": company,
+    }
+    token = authenticate()
+    response = requests.post(
+        "http://localhost:5000/clients",
+        headers={
+            "Authorization": f"Bearer {token}",
+            "Content-Type": "application/json",
+        },
+        data=json.dumps(new_client),
+    )
+    data = handle_response(response)
+    message_show_view({"Success": "Client Created"})
+    client_show_view(data)
 
 
 @app.command()
@@ -75,10 +78,76 @@ def show(id: int):
 
 
 @app.command()
+def update(
+    id: Annotated[int, typer.Option("--id", "-i", prompt=True, help="The client id")],
+    fullname: Annotated[
+        Optional[str],
+        typer.Option("--fullname", "-n", prompt=True, help="The client full name"),
+    ] = None,
+    email: Annotated[
+        Optional[str],
+        typer.Option("--email", "-e", prompt=True, help="The client email"),
+    ] = None,
+    phone: Annotated[
+        Optional[str],
+        typer.Option("--phone", "-ph", prompt=True, help="The client phone number"),
+    ] = None,
+    company: Annotated[
+        Optional[str],
+        typer.Option("--company", "-c", prompt=True, help="The company name"),
+    ] = None,
+):
+    payload = {}
+    payload["fullname"] = sanitize_fullname(fullname) if fullname else None
+    if email:
+        try:
+            email_info = validate_email(
+                email, check_deliverability=False, test_environment=True
+            )
+            payload["email"] = email_info.normalized
+        except EmailNotValidError as e:
+            message_show_view({"Error": "Email format is wrong", "Details": str(e)})
+            raise typer.Exit()
+    else:
+        payload["email"] = None
+    payload["phone"] = phone if phone else None
+    payload["company"] = company if company else None
+    update_client = {}
+    for key in payload:
+        if payload[key] is not None:
+            update_client[key] = payload[key]
+    if update_client.keys():
+        token = authenticate()
+        response = requests.put(
+            f"http://localhost:5000/clients/{id}",
+            headers={
+                "Authorization": f"Bearer {token}",
+                "Content-Type": "application/json",
+            },
+            data=json.dumps(update_client),
+        )
+        data = handle_response(response)
+        message_show_view({"Success": "Client Updated"})
+        client_show_view(data)
+    else:
+        message_show_view({"Error": "Nothing to update"})
+
+
+@app.command()
 def delete(id: int):
     token = authenticate()
+    client = requests.get(
+        f"http://localhost:5000/clients/{id}",
+        headers={"Authorization": f"Bearer {token}"},
+    )
+    data = handle_response(client)
+    client_show_view(data)
+    typer.confirm(
+        "Do you really want to delete this client? There is no going back.", abort=True
+    )
     response = requests.delete(
         f"http://localhost:5000/clients/{id}",
         headers={"Authorization": f"Bearer {token}"},
     )
-    print(handle_response(response))
+    data = handle_response(response)
+    message_show_view(data)
